Update image paths and improve HTML structure and semantics
Updated the path of the background image in the style.css file to correctly link to the new location in the media/images directory. Cleaned up unnecessary tags in the semantic-html.html file and added an 'alt' attribute to enhance the accessibility of the site. Also, a title was added to the nyc-blog.html file and the lang attribute was added to the html tag in vacation-spots.html to enhance SEO. The 'alt' attribute of the main image of the vacation-spots.html file was also filled out to improve accessibility.